refactor: :recycle: reorganize server initialization and route handling
