+++ backend/src/routes/index.ts
@@ -1,12 +1,10 @@
-import { RequestHandler, Router } from 'express';
+import { Router } from 'express';
 import { RideController } from '../controllers/RideController';
 
 const router = Router();
 
 router.post('/ride/estimate', RideController.estimateRide);
 router.patch('/ride/confirm', RideController.confirmRide);
-router.get(
-  '/ride/:customer_id',
-  RideController.getRideHistory as RequestHandler
-);
+router.get('/ride/:customer_id', RideController.getRideHistory);
+
 export default router;
